requirements.txt
customtkinter>=5.2.0
mss>=9.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
pynput>=1.7.6
requests>=2.31.0
deep-translator>=1.11.4
lingodotdev>=1.3.0
